reorganizing redux stuff
