+++ generator/src/components/AddGasSpecies.js
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { connect } from "react-redux";
-import { addGasSpecies } from "../actions";
+import { addGasSpecies } from "../redux/actions";
 
 function AddGasSpecies(props) {
