+++ generator/src/components/AddProperty.js
@@ -1,7 +1,7 @@
 import React, { useState } from "react";
 import { connect } from "react-redux";
 import Dropdown from "react-bootstrap/Dropdown";
-import { addProperty } from "../actions";
+import { addProperty } from "../redux/actions";
 
 function AddProperty(props) {
