+++ generator/src/redux/reducers/index.js
