+++ generator/src/redux/selectors/index.js
